Consolidate create_test_agent into cli/testing.py
cli/testing.py already housed sibling helpers (create_test_agent_state,
create_agent_with_events_dir). The standalone agents/testing.py module
was redundant -- move create_test_agent next to its peers and drop the
new module.
